requirements.txt
streamlit
openai
replicate
google-generativeai
transformers
python-dotenv